prettier + readme

+++ actions/clean-old.js
@@ -1,11 +1,11 @@
-const fs = require("fs");
+const fs = require('fs');
 const {
   backupDir,
   viewsBackupDir,
   materializedViewsBackupDir,
   proceduresBackupDir,
-} = require("./constants");
-const { resetLog } = require("./logger");
+} = require('./constants');
+const { resetLog } = require('./logger');
 
 const cleanOldBackups = () => {
   // Apaga a pasta de backup de views
@@ -15,8 +15,7 @@ const cleanOldBackups = () => {
     fs.rmSync(materializedViewsBackupDir, { recursive: true });
 
   // Apaga a pasta de backup de procedures
-  if (fs.existsSync(proceduresBackupDir))
-    fs.rmSync(proceduresBackupDir, { recursive: true });
+  if (fs.existsSync(proceduresBackupDir)) fs.rmSync(proceduresBackupDir, { recursive: true });
 };
 
 const createBackupDir = () => {
@@ -29,8 +28,7 @@ const createBackupDir = () => {
     fs.mkdirSync(materializedViewsBackupDir, { recursive: true });
 
   // Cria a pasta de backup de procedures
-  if (!fs.existsSync(proceduresBackupDir))
-    fs.mkdirSync(proceduresBackupDir, { recursive: true });
+  if (!fs.existsSync(proceduresBackupDir)) fs.mkdirSync(proceduresBackupDir, { recursive: true });
 };
 
 const resetAllBackups = () => {
